move all files to deprecated
