--- Contact.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import "../node_modules/bootstrap/dist/css/bootstrap.min.css"
-const Contact=()=>{
-      return <h1 className='text-center'>this is contact   s page</h1>
-}
-export default Contact;
