--- Error.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-const Error=()=>{
-      return (
-            <>
-            <div>
-       
-      <h1>oops this is error page</h1>
-      <NavLink to="/">Go back</NavLink>
-      </div>
-      </>
-      )
-}
-export default Error;
