+++ deprecated/Service.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+const Service=(props)=>{
+      return <h1>this is {props.name}  page</h1>
+}
+export default Service;
