+++ deprecated/Sresult.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+const Sresult=(props)=>{
+      const img=`https://source.unsplash.com/600x400?${props.name} `;
+      return(
+            <>
+            <div>
+            <img src={img} alt='search'></img>
+            </div>
+            </>
+      )
+      
+}
+export default Sresult;
