+++ deprecated/depth-first-s.py
