+++ deprecated/favicon.ico
