+++ deprecated/index.js
