+++ deprecated/postcss.config.js
@@ -1,0 +1,4 @@
+const tailwindcss = require("tailwindcss");
+module.exports = {
+  plugins: [tailwindcss("./tailwind.config.js"), require("autoprefixer")],
+};
