--- postcss.config.js
@@ -1,4 +1,0 @@
-const tailwindcss = require("tailwindcss");
-module.exports = {
-  plugins: [tailwindcss("./tailwind.config.js"), require("autoprefixer")],
-};
