--- style.css
@@ -1,28 +1,0 @@
-*{
-    box-sizing:border-box;
-    padding: 0;
-    margin: 0;
-}
-body{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-.emoji
-{
-    font-size: 10rem;
-    cursor: auto;
-    user-select: none;
-    
-}
-.closed{
-    display: none;
-}
-.open{
-    display: none;
-}
-.active{
-    display: block;
-}
